feat: major changes

+++ src/routes/renderer/+page.svelte
@@ -22,21 +22,26 @@
 		image_src: "",
 		colour: "",
 		customisable_parameters: [],
+		tested_on: [],
 	};
 
 	let loading = true;
+	let defaultTheme = false;
 
 	let uninstallerListen: UnlistenFn;
 
 	onMount(async () => {
 		try {
 			theme = await invoke("now_serving");
-			console.log(theme);
 			loading = false;
 			const documents = await documentDir();
 			const appDir = await join(documents, "Zefir's Flashy Cooler");
 			const preview_img = appDir + "\\" + theme.fs_name + "\\preview.jpg";
 			theme.image_src = convertFileSrc(preview_img);
+			if (theme.fs_name === "__DEFAULT__") {
+				defaultTheme = true;
+				theme.image_src = "/images/default.png";
+			}
 		} catch (e) {
 			console.log(e);
 		}
@@ -76,22 +81,24 @@
 		</div>
 		<div class="preview-card">
 			<img src={theme.image_src} alt="theme preview" class="preview-image" />
-			<div class="btn-group">
-				<button
-					class="btn btn-outline-danger"
-					on:click={() => invoke("uninstall_theme", { fsName: theme.fs_name })}
-				>
-					<svg
-						class="svg-danger"
-						xmlns="http://www.w3.org/2000/svg"
-						height="1em"
-						viewBox="0 0 448 512"
-						><!--! Font Awesome Free 6.4.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc. --><path
-							d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"
-						/></svg
-					>Delete from drive
-				</button>
-			</div>
+			{#if !defaultTheme}
+				<div class="btn-group">
+					<button
+						class="btn btn-outline-danger"
+						on:click={() => invoke("uninstall_theme", { fsName: theme.fs_name })}
+					>
+						<svg
+							class="svg-danger"
+							xmlns="http://www.w3.org/2000/svg"
+							height="1em"
+							viewBox="0 0 448 512"
+							><!--! Font Awesome Free 6.4.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc. --><path
+								d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"
+							/></svg
+						>Delete from drive
+					</button>
+				</div>
+			{/if}
 		</div>
 	</div>
 	{#if theme.customisable_parameters.length > 0}
